requirements.txt
pyyaml
appdirs
peewee